[BABY-12] applying go.mod WIP
