+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/Junyong-Suh/babystep-api-gateway-go
+
+go 1.12
+
+require github.com/gorilla/mux v1.7.1
