+++ handlers/root_handler_test.go
@@ -1,26 +1,23 @@
-package main
+package handlers
 
 import (
     "net/http"
     "net/http/httptest"
     "testing"
     h "../handlers"
-    "fmt"
-    "io/ioutil"
-    "strings"
 )
 
-func TestHealthCheckHandler(t *testing.T) {
+func TestRootHandler(t *testing.T) {
     // Create a request to pass to our handler. We don't have any query parameters for now, so we'll
     // pass 'nil' as the third parameter.
-    req, err := http.NewRequest("GET", "/health", nil)
+    req, err := http.NewRequest("GET", "/", nil)
     if err != nil {
         t.Fatal(err)
     }
 
     // We create a ResponseRecorder (which satisfies http.ResponseWriter) to record the response.
     rr := httptest.NewRecorder()
-    handler := http.HandlerFunc(h.HealthcheckHandler)
+    handler := http.HandlerFunc(h.RootHandler)
 
     // Our handlers satisfy http.Handler, so we can call their ServeHTTP method
     // directly and pass in our Request and ResponseRecorder.
@@ -33,9 +30,7 @@ func TestHealthCheckHandler(t *testing.T) {
     }
 
     // Check the response body is what we expect.
-    version, _ := ioutil.ReadFile("./VERSION")
-    v := strings.TrimSpace(string(version))
-    expected := fmt.Sprintf("{\"version\": \"%s\", \"service\": \"babystep-api-gateway\"}", v)
+    expected := `Hello`
     if rr.Body.String() != expected {
         t.Errorf("handler returned unexpected body: got %v want %v",
             rr.Body.String(), expected)
